Xtensa: Fix scalar ae_f/ae_p HiFi type sizes to match memory footprint
The scalar (x1) HiFi ae_ types were all defined as int64_t, giving
sizeof(ae_f32)=8, sizeof(ae_f16)=8, sizeof(ae_f24)=8, sizeof(ae_p24f)=8,
sizeof(ae_p16s)=8.  This breaks any code that uses sizeof() or pointer
arithmetic on these types, because pointer increments and memcpy sizes
come out 2-4x too large.

In Cadence's TIE type system, scalar types carry their natural memory
footprint as sizeof(), while vector/packed types fill the 64-bit AEDR
register.  The correct mapping is:

  ae_int32  -> int32_t  (already correct)
  ae_int16  -> int16_t  (already correct)
  ae_f32    -> int32_t  (scalar 32-bit fixed-point; sizeof == 4)
  ae_f16    -> int16_t  (scalar 16-bit fixed-point; sizeof == 2)
  ae_f24    -> int32_t  (scalar 24-bit fixed-point; sizeof == 4)
  ae_p24f   -> int32_t  (scalar 24-bit fractional;  sizeof == 4)
  ae_p16s   -> int16_t  (scalar 16-bit fractional;  sizeof == 2)

Vector/packed types (ae_f32x2, ae_f24x2, ae_f16x4, ae_p24x2f,
ae_p16x2s, ae_int32x2, ae_int16x4) remain int64_t as they occupy a
full 64-bit AEDR register.

The AE_L32_XP and similar macros cast through (void**) so the pointer
element type does not affect intrinsic call sites.  Assignment of an
int64_t intrinsic result to an int32_t ae_f32 lvalue truncates the
upper 32 bits, which are zero/sign-extension of the actual 32-bit value,
so no information is lost.

Tested on Intel Tiger Lake (CAVS 2.5) with SOF firmware: all HiFi3
audio modules (volume, mixin/mixout, copier, eq_fir, drc, asrc, aria,
dcblock) pass check-playback and check-capture with all eight
CONFIG_COMP_HIFI modules at HIFI_MAX.

+++ clang/lib/Headers/xtensahifiintrin.h
@@ -14,21 +14,22 @@
 
 #include <stdint.h>
 
-/* All HiFi ae_ types are 64-bit AEDR register values.
- * In Cadence's compiler they are opaque TIE types with C-cast support.
- * We represent them as int64_t since C allows scalar casts between
- * int64_t and int32_t/int16_t, matching Cadence's cast semantics.
+/* All HiFi ae_ types map to 64-bit AEDR register values in hardware.
+ * In Cadence's TIE type system, scalar (x1) types carry their memory
+ * footprint as sizeof, while vector (x2/x4) types fill the 64-bit register.
+ * Match that convention so that pointer arithmetic on ae_f32 ptr, ae_f16 ptr
+ * etc. advances by the correct element size, just as ae_int32/ae_int16 do.
  * The builtins use v2i32/v4i16; macros bridge via __ae_v2i32/__ae_v4i16. */
 typedef int64_t ae_int32x2;
 typedef int64_t ae_int16x4;
 typedef int64_t ae_f64;
 typedef int32_t ae_int32;
 typedef int16_t ae_int16;
-typedef int64_t ae_f32;
-typedef int64_t ae_f32x2;
-typedef int64_t ae_f24x2;
-typedef int64_t ae_f16;
-typedef int64_t ae_f16x4;
+typedef int32_t ae_f32;   /* scalar 32-bit fixed-point; sizeof == 4 */
+typedef int64_t ae_f32x2; /* 2x32-bit packed; sizeof == 8 */
+typedef int64_t ae_f24x2; /* 2x24-bit packed; sizeof == 8 */
+typedef int16_t ae_f16;   /* scalar 16-bit fixed-point; sizeof == 2 */
+typedef int64_t ae_f16x4; /* 4x16-bit packed; sizeof == 8 */
 typedef int64_t ae_int64;
 typedef int64_t ae_valign;
 typedef int64_t ae_q56s;
@@ -36,11 +37,11 @@ typedef int64_t ae_q56s;
 /* Legacy/alternate names used by SOF */
 typedef int64_t ae_p24x2f;
 typedef int64_t ae_p16x2s;
-typedef int64_t ae_p24f;
-typedef int64_t ae_p16s;
+typedef int32_t ae_p24f;  /* scalar 24-bit fractional; sizeof == 4 */
+typedef int16_t ae_p16s;  /* scalar 16-bit fractional; sizeof == 2 */
 typedef int64_t ae_p24x2s;
 typedef int64_t ae_int24x2;
-typedef int64_t ae_f24;
+typedef int32_t ae_f24;   /* scalar 24-bit fixed-point; sizeof == 4 */
 
 /* HiFi5 128-bit types — pairs of 64-bit AE register values.
  * These are used by HiFi5 "x2" instructions (ae_l32x2x2, etc.) which
